backend/backend_services/profile/app/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyjwt
gunicorn
requests
dotmap